+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog], and this project adheres to
 [Semantic Versioning].
 
+## [0.9.1] - 2020-07-01
+
+### Added
+* Added `Mat3 * Vec3A` implementation.
+* Added `Vec3A` benches.
+
+### Changed
+* Some documentation improvements around the new `Vec3A` type.
+
 ## [0.9.0] - 2020-06-28
 
 ### Added
@@ -186,7 +195,8 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 [Keep a Changelog]: https://keepachangelog.com/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
-[Unreleased]: https://github.com/bitshifter/glam-rs/compare/0.9.0...HEAD
+[Unreleased]: https://github.com/bitshifter/glam-rs/compare/0.9.1...HEAD
+[0.9.1]: https://github.com/bitshifter/glam-rs/compare/0.9.0...0.9.1
 [0.9.0]: https://github.com/bitshifter/glam-rs/compare/0.8.7...0.9.0
 [0.8.7]: https://github.com/bitshifter/glam-rs/compare/0.8.6...0.8.7
 [0.8.6]: https://github.com/bitshifter/glam-rs/compare/0.8.5...0.8.6